requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
openai==0.28.1
gTTS==2.4.0
whisper==1.1.10
pydub==0.25.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
numpy==1.24.4
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
flask-sqlalchemy==3.1.1
boto3==1.28.57
python-dotenv==1.0.0
gunicorn==21.2.0
google-cloud-storage==2.10.0